added confirm functionality to delete buttons, moved photoshoot onClick to details page to section tag holding photoshoot info because delete confirm was activating the details onClick when canceled

+++ src/components/client/ClientDetails.js
@@ -13,8 +13,11 @@ const ClientDetails = props => {
     };
 
     const handleDelete = () => {
-        ApiManager.delete("clients", clientId)
-            .then(() => props.history.push('/clients'))
+        const confirmed = window.confirm("are you sure?")
+        if (confirmed === true) {
+            ApiManager.delete("clients", clientId)
+                .then(() => props.history.push('/clients'))
+        }
     }
 
     useEffect(() => {
